refactor!: Move Base AuthClient Types to authclient.ts
