+++ src/auth/externalAccountAuthorizedUserClient.ts
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {AuthClient, AuthClientOptions} from './authclient';
-import {Headers} from './oauth2client';
+import {AuthClient, AuthClientOptions, Headers} from './authclient';
 import {
   ClientAuthentication,
   getErrorFromOAuthErrorResponse,
