+++ src/auth/oauth2client.ts
@@ -25,7 +25,7 @@ import * as formatEcdsa from 'ecdsa-sig-formatter';
 import {createCrypto, JwkCertificate, hasBrowserCrypto} from '../crypto/crypto';
 import {BodyResponseCallback} from '../transporters';
 
-import {AuthClient, AuthClientOptions} from './authclient';
+import {AuthClient, AuthClientOptions, GetAccessTokenResponse, Headers} from './authclient';
 import {CredentialRequest, Credentials} from './credentials';
 import {LoginTicket, TokenPayload} from './loginticket';
 /**
@@ -51,10 +51,6 @@ export interface Certificates {
 }
 
 export interface PublicKeys {
-  [index: string]: string;
-}
-
-export interface Headers {
   [index: string]: string;
 }
 
@@ -334,11 +330,6 @@ export interface GetAccessTokenCallback {
     token?: string | null,
     res?: GaxiosResponse | null
   ): void;
-}
-
-export interface GetAccessTokenResponse {
-  token?: string | null;
-  res?: GaxiosResponse | null;
 }
 
 export interface RefreshAccessTokenCallback {
